* [Platform] All DAO::delete() calls will now properly bailout if the ID list they're given is empty.  Normally this is not necessary, but there have been a few situations where users have cached URLs in their local browser that attempt to repeat actions (like deleting worklists or tasks).

+++ api/DAO.php
@@ -593,6 +593,9 @@ class DAO_Translation extends DevblocksORMHelper {
 		if(!is_array($ids)) $ids = array($ids);
 		$db = DevblocksPlatform::getDatabaseService();
 		
+		if(empty($ids))
+			return;
+		
 		$ids_list = implode(',', $ids);
 		
 		$db->Execute(sprintf("DELETE FROM translation WHERE id IN (%s)", $ids_list));
